Fix 0.6 and 0.7 flags being mixed up
If the client is connecting via sixup. Always send 0.7 flags.
But only translate on send to use 0.6 all over the client code base.

This fixes the hook and double jump animation not being rendered
while the scoreboard is active. Because earlier it would store the
translated flags in the input history and then confuse scoreboard
with chatting.

All those kind of issues should not be fixed since it is now ensured
the client is only aware of the 0.6 flag until the very last moment
before it is being sent to the server.