debug: seems like soundworld is red at the wrong offset
When recording sound world is dumped correctly when firing the gun.
When playing the demo it seems like it starts reading one int later.
So the x coordinate is a random value. The y is the x. And the sound id
is the y. That then fails the sound id validation because the value is
too high.