Add timeout for fifo in integration tests
If the client or server crashes during the integration test run
the next write to its fifo file will be blocking and get stuck forever.

This is especially annoying for github actions because they just run 10
hours and then get terminated by github without ever reaching the end
of the integration test script that then prints the details about the
crash.