Move size inside `NetTwIntString` type, fix debug dump
The code for the debug dump of `NetTwIntString`s was not being emitted due to it being wrapped in a `NetArray`. It also did not compile because variables `aStr` and `aInts` were not defined.

The string size in `char`s is now passed as an argument to `NetTwIntString`, which is a specialization of `NetArray`, instead of wrapping the `NetTwIntString` in a `NetArray`.