Use words instead of numbers for access level commands
The new encouraged way to refer to access levels in configs is
to use the access level name as a string. The magic number
integer is still kept as an backwards compatibility alias.

And there is currently no plan to deprecate the old alias.

Here a example config

```
# old, still works
access_level ban 1

# new, does the same
access_level ban moderator

# old, still works
access_level kick 2

# new, does the same
access_level kick helper
```

This is a preparation for #10681

# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
#
# Date:      Mon Aug 18 17:34:28 2025 +0200
#
# On branch pr_accessstr
# Your branch is up to date with 'origin/pr_accessstr'.
#
# Changes to be committed:
#	modified:   data/autoexec_server.cfg
#	modified:   src/engine/shared/console.cpp
#	modified:   src/engine/shared/console.h
#