Add rcon role class
Also refer to these roles by name using a string.
Instead of using an integer.

This is a preperation for #10681.
To be able to introduce new roles at runtime and referring to
them by name. This will also allow to store command access in the role
in the future. Instead of storing the access level in the command
how it is done right now. That shift allows to have multiple roles
that can have access to a different set of commands. Right now
access level access management is limited because the different levels
are forced to inherit command access from the lower levels. Making it
impossible to create roles with different responsibilities.

This commit is just an internal refactor. Nothing user facing changes.