Change access level order
Now a higher access level means more access.
Makes sense in my head.

I assume it was done the other way around to have the magic value
zero reserved for the admin rank. And to be able to just append lower
levels at the bottom of the enum such as moderator, helper and possibly
more. But the access code readability suffers from this.
And adding more rcon access levels is currently planned to be done with
roles. That will work different. And also there we want higher numbers
to mean more power.