Revert skin name normalization
Fix skins not being loaded and favorite skins not working on case-sensitive file systems, if the name used to load the skin for the first time does not have the same capitalization as the actual file in `downloadedskins`. For example, with `add_favorite_skin bub` in the config file, this would cause the skin loading to use the name `bub`, but the actual name of the skin is `Bub`. On a case-insensitive file system, the client would successfully load the skin from `Bub.png`, but on case-sensitive file systems it would only accept `bub.png`. Either way, the skin download server would also respond with 404 Not Found if `bub.png` is requested instead of `Bub.png`.

Reintroduces that the same skins can be loaded multiple times for different capitalizations on case-insensitive file systems.