Step instead of ExecuteUpdate in mysql AddPoints
Brings the mysql version closer to the sqlite3 version.
The method `ExecuteUpdate()` is just a wrapper around `Step()`
that can be used if we are interested in the amount of
affected rows.

Since that value is not checked we can just `Step()` instead.

Another possibility would be to throw and assert or return
an error unless exactly 1 row was affected.